rewrite fragments
In this PR we addressed the issue with broken fragments merging. Historically, hiku has not supported unions/interfaces so fragments merging were ok. But after we added unions/intefaces support + fragments support, things went in the wrong direction.

Fragments merging (fields merging to be more accurate) has its own rules in graphql spec and we did not complied. So why thought ?

Hiku arhitecture was build without fragments in mind, that is - engine and denormalization modules can not work with fragments. So to avoid complex rewriting for engine and denormalization we decided to merge fragments at query parsing stage to avoid fields duplication - the most hard thing to support in engine.

Doing fragments merging at query parsing stage was a bad decision since
we do not have access to graph, hence can not know how to merge
particular fragment because merge rules for unions and interfaces are
different.

In this PR we inroducing new `QueryMerger` that merges fragments after
parsing but before execution, thus we do not need to rewrite engine much
and just make a small tweaks to denormalized.

Some other changes:
 - Refactor SplitQuery types, introduce FieldInfo and LinkInfo instead of tuples
 - Refactor GroupQuery to use FieldInfo/LinkInfo
 - Fix cache tests
 - Drop fragments hack from result.py:Proxy since we now provide proper Proxy to index for each fragment
 - Add name for Fragment, if name is None - this is an InlineFragment,
   make type_name optional, because fragments can be without type
   condition
 - Node.fragments_map only returns named fragments map